training/requirements.txt
pandas
pillow
keras
scikit-learn
matplotlib
numpy
pandas
pyyaml
kaleido
faiss-cpu
py-cpuinfo
psutil
wandb
tqdm
h5py
torch-summary
